requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
websockets==12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
celery==5.3.4
redis==5.0.1
python-magic-link==0.5.1
google-auth==2.23.4
ell-ai==1.0.2
python-dotenv==1.0.0